requirements.txt
numpy==1.26.0
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
aiohttp==3.9.1
schedule==1.2.0
selenium==4.29.0
webdriver-manager==4.0.2
sendgrid==6.11.0
fake-useragent==1.4.0
flask==3.0.0
flask-cors==4.0.0